fix: 채팅방 목록 조회 오류 수정

+++ src/main/java/com/planu/group_meeting/chat/dto/response/ChatRoomResponse.java
@@ -15,5 +15,5 @@ public class ChatRoomResponse {
     private String lastChat;
     private String lastChatDate;
     private String lastChatTime;
-    private Long unreadChats;
+    private Integer unreadChats;
 }
